Se creo proyecto de Unit Test. Se cambio logica en deletePet en PetService para el test.

+++ Petfy.Domain/Services/PetService.cs
@@ -99,6 +99,21 @@ namespace Petfy.Domain.Services
 
         public bool DeletePet(int id)
         {
+            if (id <= 0)
+            {
+                return false;
+            }
+
+            var pet = _petRepository.GetById(id);
+            if (pet == null)
+            {
+                throw new ApplicationException("Pet not found");
+            }
+            if (pet.Owner == null )
+            {
+                throw new ApplicationException("Pet cant be deleted because it has an owner");
+            }
+
             try
             {
                 return _petRepository.DeletePet(id);
